requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
discord.py==2.3.2
Flask==3.0.3
frozenlist==1.4.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
numpy==1.26.4
Werkzeug==3.0.3
yarl==1.9.4
